% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateSS.R
\name{dmvnorm_own}
\alias{dmvnorm_own}
\title{Density of multivariate normal distribution}
\usage{
dmvnorm_own(y, mu = rep(0, length(y)), sigma = diag(1, length(y)), tau = 0.1)
}
\arguments{
\item{y}{vector of quantiles}

\item{mu}{mean vector, default is \code{rep(0, length(y))}}

\item{sigma}{covariance matrix, default is \code{diag(1, length(y))}}

\item{tau}{constant multiplier (positive) for the covariance matrix, default is 0.1}
}
\value{
It returns the density function for the multivariate normal distribution with mean vector \eqn{\mu}, covariance matrix \eqn{\tau \times \Sigma}
}
\description{
Density of multivariate normal distribution
}
\examples{
dmvnorm_own(y = c(0,0))
dmvnorm_own(y = c(0,0), mu = c(1,1))
dmvnorm_own(y = c(0,0), mu = c(1,1), sigma = diag(2, 2), tau = 1)
}
