% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESS.R
\name{ESS}
\alias{ESS}
\title{Title Elliptical Slice Sampling}
\usage{
ESS(x, mu, sigma, log.L, niter)
}
\arguments{
\item{x}{current state}

\item{mu}{mean of Gaussian process}

\item{sigma}{sd of Gaussian process}

\item{log.L}{log-likelihood function}

\item{niter}{maximum number of iterations in one update}
}
\value{
It will return the new state x'
}
\description{
Title Elliptical Slice Sampling
}
\examples{
set.seed(12345)
x <- seq(1, 1e-3, -0.1)
ff <- f_data(x)
y <- ff + rnorm(length(x), 0, sd(ff)/10)
training_data = cbind(y, x)




}
