% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForExtension.R
\name{inv_and_logdet}
\alias{inv_and_logdet}
\title{Calculate the inverse and log(determinant) of symmetric matrix}
\usage{
inv_and_logdet(Sigma)
}
\arguments{
\item{Sigma}{is a symmetric matrix whose inverse and determinant are to be calculated}
}
\value{
A list containing the inverse and the log(determinant) of the matrix
}
\description{
Calculate the inverse and log(determinant) of symmetric matrix
}
\examples{
mat <- diag(c(2, 4))
inv_and_logdet(mat)
}
