% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateSS.R
\name{mySolve}
\alias{mySolve}
\title{Inversion of a symmetric matrix}
\usage{
mySolve(A)
}
\arguments{
\item{A}{a symmetric matrix}
}
\value{
A list containing
\item{inv}{the inverse}
\item{det}{the determinant}
\item{sqrtDeco}{the square root matrix \eqn{Q} of \eqn{A}, such that it holds \eqn{QQ=A}}
}
\description{
Inversion of a symmetric matrix
}
\examples{
A = diag(2, 2)
mySolve(A)
}
