% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESS.R
\name{ESS}
\alias{ESS}
\title{Title Elliptical Slice Sampling}
\usage{
ESS(x, mu, sigma, log.L)
}
\arguments{
\item{x}{current state}

\item{mu}{mean of Gaussian process}

\item{sigma}{sd of Gaussian process}

\item{log.L}{log-likelihood function}
}
\value{
new state x'
}
\description{
Title Elliptical Slice Sampling
}
