% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateSS.R
\name{mySolve}
\alias{mySolve}
\title{Inversion of a symmetric matrix}
\usage{
mySolve(A)
}
\arguments{
\item{A}{a symmetric matrix}
}
\value{
A list containing the inverse, determinant and the square root matrix $Q$ of $A$, that is, it holds $QQ=A$
}
\description{
Inversion of a symmetric matrix
}
\examples{
A = diag(2, 2)
mySolve(A)

}
