% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateMH.R
\name{SurrogateMH}
\alias{SurrogateMH}
\title{Surrogate data model - Metropolis Hastings}
\usage{
SurrogateMH(theta, f, data, p)
}
\arguments{
\item{theta}{hyperparameter of interest}

\item{f}{latent variable}

\item{data}{the data in hand}

\item{p}{the prior distribution}
}
\value{
updated theta and f
}
\description{
Surrogate data model - Metropolis Hastings
}
