% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateSS.R
\name{SurrogateSS}
\alias{SurrogateSS}
\title{Surrogate data model - Slice sampling}
\usage{
SurrogateSS(theta, f, sigma, data, l, p)
}
\arguments{
\item{theta}{hyperparameter of interest}

\item{f}{latent variable}

\item{sigma}{scale parameter}

\item{data}{the data in hand}

\item{l}{conditional likelihood function}

\item{p}{prior distribution of the covariance hyperparameter}
}
\value{
updated theta and f
}
\description{
Surrogate data model - Slice sampling
}
